removed the unneededcode
