--- phone-radio/src/utils/baseUrl.ts
@@ -1,7 +1,0 @@
-import type { Request } from "express";
-
-export function baseUrl(req: Request): string {
-  const proto = req.header("x-forwarded-proto") ?? req.protocol;
-  const host = req.header("x-forwarded-host") ?? req.header("host");
-  return `${proto}://${host}`;
-}
